Radix sort now working with literal cast inference.

+++ library/thrust_wrappers/index_sequence.h
@@ -5,6 +5,6 @@
 struct index_sequence : public iterator_sequence<thrust::counting_iterator<long> >
 {
   __host__ __device__ index_sequence(long length) :
-  iterator_sequence<thrust::counting_iterator<long> >(thrust::counting_iterator<int>(0), length) {}
+  iterator_sequence<thrust::counting_iterator<long> >(thrust::counting_iterator<long>(0), length) {}
 };
